DTPS/requirements.txt
fastapi==0.101.1
jinja2==3.1.2
python-multipart==0.0.6
influxdb_client==1.37.0
joblib==1.3.2
kubernetes==26.1.0
marshmallow==3.20.1
matplotlib==3.7.1
numpy==1.25.0
paho_mqtt==1.6.1
pydantic==2.2.0
pyDataverse==0.3.1
pyDataverse==0.3.1
pymongo==4.4.1
Requests==2.31.0
scipy==1.11.2
statsmodels==0.14.0
uvicorn==0.23.2
watchdog==3.0.0
